Carve vixen-primitives and vixen-runtime out of vix-core
Split the monolith into the V8/Node shape the language wants:

- vix-core (the language): parses, compiles, lowers, and executes, but ships an
  empty primitive dispatcher and an empty prelude — inert alone. It owns the
  primitive *binding descriptors* (ids, request types, surface decls +
  synth_descriptor/synth_shape, FETCH_DECL/OBSERVE_DECL) that the compiler needs
  to lower a call, but none of the implementations. binding::builtin_bindings now
  harvests from runtime::builtin_primitive_surfaces (language data), not from the
  primitive impls.

- vixen-primitives (the batteries): the RawPrimitive impls
  (decode/fetch/observe/tree-read), the typed authoring layer
  (Primitive/TypedAdapter), and the pure-vix stdlib — now one .vix file per
  prelude function under stdlib/, assembled into PRELUDE_SOURCES.

- vixen-runtime (the runnable system): registers the builtins into the engine
  (builtin_primitives/default_primitive_dispatcher/install_builtins), supplies
  the stdlib prelude (default_config), hosts the ratchet/budget harness + the
  vix-budget-child bin, and re-exports the whole surface.

vix-core dev-depends on the vixen crates (a dev-only cycle Cargo permits) so its
own tests can drive the runnable system. All four crate libs compile;
vixen-primitives tests pass (incl. the typed-adapter milestone tests that assert
the synthesized descriptors are byte-identical to the pre-split hand-written
ones). vix-core/vixen-runtime test migration follows.
