+++ vix-core/src/runtime/mod.rs
@@ -1,37 +1,31 @@
 //! Runtime ownership boundaries for the new Vix compiler path.
+//!
+//! This is the language's execution engine and the *raw* primitive protocol
+//! (`RawPrimitive`, `PrimitiveDispatcher`, `EffectCtx`, the descriptor types in
+//! [`primitive_bindings`]). It ships no concrete primitives — those, and the
+//! typed convenience layer they are written against, live in `vixen-primitives`
+//! and are installed by `vixen-runtime`.
 
 mod abi;
 mod blob_persistence;
-mod decode_primitive;
 mod error;
-mod fetch_primitive;
 mod fixture;
 mod identity;
 mod model;
 mod observe;
-mod observe_primitive;
 mod primitive;
 mod primitive_bindings;
-mod primitive_value_decode;
 mod scheduler;
 mod store;
-mod tree_read_primitive;
-mod typed_primitive;
 
 pub use abi::*;
 pub use blob_persistence::*;
-pub use decode_primitive::*;
 pub use error::*;
-pub use fetch_primitive::*;
 pub use fixture::*;
 pub use identity::*;
 pub use model::*;
 pub use observe::*;
-pub use observe_primitive::*;
 pub use primitive::*;
 pub use primitive_bindings::*;
-pub use primitive_value_decode::*;
 pub use scheduler::*;
 pub use store::*;
-pub use tree_read_primitive::*;
-pub use typed_primitive::*;
