+++ vixen-primitives/src/stdlib/json_decode.vix
@@ -1,0 +1,3 @@
+fn json_decode<T>(text: String) -> T {
+    decode(text, Format::Json)
+}
