+++ vixen-primitives/src/stdlib/mode.vix
@@ -1,0 +1,4 @@
+enum Mode {
+    Observe,
+    Refresh,
+}
