+++ vixen-primitives/src/stdlib/refresh.vix
@@ -1,0 +1,3 @@
+fn refresh<Origin>(origin: Origin) -> Blob {
+    observe(origin, Mode::Refresh)
+}
