+++ vixen-primitives/src/stdlib/toml_decode.vix
@@ -1,0 +1,3 @@
+fn toml_decode<T>(text: String) -> T {
+    decode(text, Format::Toml)
+}
