+++ vixen-primitives/src/stdlib/try_json_decode.vix
@@ -1,0 +1,3 @@
+fn try_json_decode<T>(text: String) -> Result<T, DecodeError> {
+    try_decode(text, Format::Json)
+}
